[ADD] stock_transport: modified fleet category list, fleet form, inventory batch
- Modified fleet category list by adding two fields to it. Two fields are
max_weight and max_volume related to that particular category.

- Modified fleet form view , changed display name of category. Modified it as
name of category, max_volume and max_weight concatenated with category name in
simple bracket.

-Modified form view of inventory batch. Added Many2one fields such as dock,
vehicle and vehicle category to the model by inheriting it. Also added this
fields to the form view of batch. Also added total_weight and total_volume to
the model and form view. Displayed progressbar for percentage of weight and
volume added to batch with respect to capacity of vehicle.

- Added graph view for batch transfer. Grouped them with respect to category and
scheduled date.